+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+## [v0.57.2] - 2026-05-17
+
+### Fixed
+- `dvt wezterm parse -o json`: use `json.Marshal` instead of `yaml.Marshal` in `yamlConfigToJSON` to produce valid JSON output ([#435](https://github.com/rmkohlman/devopsmaestro/issues/435))
+
+---
+
 ## [v0.105.3] - 2026-04-27
 
 ### Fixed